improve mixed precision and fix FA. Add prompt dropout for training etc

+++ examples/animatediff/ad/modules/diffusionmodules/util.py
@@ -104,8 +104,10 @@ class SiLU(nn.Cell):
         self.sigmoid = ops.Sigmoid()
 
     def construct(self, x):
+        # TODO: test use fp16 instead
         # force sigmoid to use fp32
-        return x * self.sigmoid(x.astype(ms.float32)).astype(x.dtype)
+        # return x * self.sigmoid(x.astype(ms.float32)).astype(x.dtype)
+        return x * self.sigmoid(x)
 
 
 class GroupNorm32(nn.GroupNorm):
